requirements.txt
flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
jinja2==3.0.2
markupsafe==2.0.1
pandas==1.3.4
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
werkzeug==2.0.2
